ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
supabase==2.9.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.28.1
python-multipart==0.0.9
